Move page chunking to PageAllocator
