+++ BepInEx.IL2CPP/Hook/Allocator/LinuxPageAllocator.cs
@@ -23,21 +23,24 @@ namespace BepInEx.IL2CPP.Allocator
 				throw new NotImplementedException();
 			}
 		}
-		
+
 		static class Unix
 		{
 			public delegate IntPtr mmapDelegate(IntPtr addr, UIntPtr length, int prot, int flags, int fd, int offset);
+
 			[DynDllImport("mmap")]
 			public static mmapDelegate mmap;
 
 			public delegate int munmapDelegate(IntPtr addr, UIntPtr length);
+
 			[DynDllImport("munmap")]
 			public static munmapDelegate munmap;
-			
+
 			public delegate IntPtr fdopenDelegate(int fd, string mode);
+
 			[DynDllImport("libc")]
 			public static fdopenDelegate fdopen;
-			
+
 			static Unix()
 			{
 				typeof(Unix).ResolveDynDllImports(new Dictionary<string, List<DynDllMapping>>
