+++ BepInEx.IL2CPP/Hook/Allocator/WindowsPageAllocator.cs
@@ -10,15 +10,7 @@ namespace BepInEx.IL2CPP.Allocator
 	/// </summary>
 	internal class WindowsPageAllocator : PageAllocator
 	{
-		private readonly List<PageChunk> allocatedChunks = new List<PageChunk>();
-
-		/// <summary>
-		///     Allocates a single 64k chunk of memory near the given address
-		/// </summary>
-		/// <param name="hint">Address near which to attempt allocate the chunk</param>
-		/// <returns>Allocated chunk</returns>
-		/// <exception cref="Win32Exception">Allocation failed</exception>
-		private PageChunk AllocateChunk(IntPtr hint)
+		protected override IntPtr AllocateChunk(IntPtr hint)
 		{
 			while (true)
 			{
@@ -51,70 +43,15 @@ namespace BepInEx.IL2CPP.Allocator
 				throw new Win32Exception(error);
 			}
 
-			var result = new PageChunk
-			{
-				BaseAddress = chunk
-			};
-			allocatedChunks.Add(result);
-			return result;
-		}
-
-		private IntPtr AllocPage(IntPtr hint)
-		{
-			foreach (var allocatedChunk in allocatedChunks)
-			{
-				// Small shortcut to speed up page lookup
-				if (allocatedChunk.UsedPages == PAGES_PER_UNIT)
-					continue;
-				for (var i = 0; i < allocatedChunk.Pages.Length; i++)
-				{
-					if (allocatedChunk.Pages[i])
-						continue;
-					var pageAddr = allocatedChunk.GetPage(i);
-					if (!IsInRelJmpRange(hint, pageAddr))
-						continue;
-					allocatedChunk.Pages[i] = true;
-					allocatedChunk.UsedPages++;
-					return pageAddr;
-				}
-			}
-
-			var chunk = AllocateChunk(hint);
-			chunk.Pages[0] = true;
-			chunk.UsedPages++;
-			return chunk.BaseAddress;
+			return chunk;
 		}
 
 		public override IntPtr Allocate(IntPtr hint)
 		{
-			var pageAddress = AllocPage(hint);
+			var pageAddress = base.Allocate(hint);
 			if (WinApi.VirtualAlloc(pageAddress, (UIntPtr)PAGE_SIZE, WinApi.AllocationType.MEM_COMMIT, WinApi.ProtectConstant.PAGE_READWRITE) == IntPtr.Zero)
 				throw new Win32Exception(Marshal.GetLastWin32Error());
 			return pageAddress;
-		}
-
-		public override void Free(IntPtr page)
-		{
-			foreach (var allocatedChunk in allocatedChunks)
-			{
-				long index = (page.ToInt64() - allocatedChunk.BaseAddress.ToInt64()) / PAGE_SIZE;
-				if (index < 0 || index > PAGES_PER_UNIT)
-					continue;
-				allocatedChunk.Pages[index] = false;
-				return;
-			}
-		}
-
-		private class PageChunk
-		{
-			public readonly bool[] Pages = new bool[PAGES_PER_UNIT];
-			public IntPtr BaseAddress;
-			public int UsedPages;
-
-			public IntPtr GetPage(int index)
-			{
-				return BaseAddress + index * PAGE_SIZE;
-			}
 		}
 
 		private static class WinApi
